Revert "feat: article 패키지 오타수정"
This reverts commit aaf5a575e07740215eae3a7cfa24a96ca4af9496.
