+++ src/main/java/com/happiday/Happi_Day/domain/entity/aticle/UserArticleLikes.java
@@ -1,8 +1,7 @@
-package com.happiday.Happi_Day.domain.entity.article;
+package com.happiday.Happi_Day.domain.entity.aticle;
 
-import com.happiday.Happi_Day.domain.entity.BaseEntity;
+import com.happiday.Happi_Day.domain.entity.aticle.Article;
 import jakarta.persistence.*;
-import jakarta.validation.constraints.NotNull;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -14,24 +13,19 @@ import lombok.experimental.SuperBuilder;
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor
 @SuperBuilder(toBuilder = true)
-@Table(name = "comment")
-public class Comment extends BaseEntity {
+@Table(name = "user_article_likes")
+public class UserArticleLikes {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
-
-    // 게시글 id
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name="article_id")
-    private Article article;
 
     // 유저 id
 //    @ManyToOne(fetch = FetchType.LAZY)
 //    @JoinColumn(name="user_id")
 //    private User user;
 
-    @NotNull
-    @Column(columnDefinition = "TEXT")
-    private String content;
-
+    // 게시글 id
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name="article_id")
+    private Article article;
 }
