+++ src/main/java/com/happiday/Happi_Day/domain/entity/aticle/UserArticleScrap.java
@@ -1,5 +1,6 @@
-package com.happiday.Happi_Day.domain.entity.article;
+package com.happiday.Happi_Day.domain.entity.aticle;
 
+import com.happiday.Happi_Day.domain.entity.aticle.Article;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
